test(my-space): #4056 — split MissingInfoModal refresh tests into a dedicated file
Extract the router.refresh RSC-refresh suite into MissingInfoModal.refresh.test.tsx
to bring the original test file under the 400-line threshold (rule-2.5), and drop two
paraphrasing comments (rule-2.17).
